requirements.txt
drf_yasg==1.20.0
Django==3.2.5
djangorestframework==3.12.4
django-rest-auth~=0.9.5